Don't detect spark version in production code. It does not bring much benefits as we still don't deploy selectively in runtime, so we need spark 2 artifacts to be sent even to spark3 cluster
