+++ cdap-app-fabric/src/main/java/io/cdap/cdap/internal/app/spark/SparkCompatReader.java
@@ -53,12 +53,7 @@ public class SparkCompatReader {
     compatStr = compatStr == null ? cConf.get(Constants.AppFabric.SPARK_COMPAT) : compatStr;
 
     if (compatStr == null) {
-      try {
-        return getCurrentSparkCompat();
-      } catch (Throwable e) {
-        LOG.info("Can't detect spark version({}), using default {}", e, SparkCompat.SPARK2_2_11);
-        return SparkCompat.SPARK2_2_11;
-      }
+      return SparkCompat.SPARK2_2_11;
     }
 
     for (SparkCompat sparkCompat : SparkCompat.values()) {
@@ -74,18 +69,5 @@ public class SparkCompatReader {
 
     throw new IllegalArgumentException(
       String.format("Invalid SparkCompat version '%s'. Must be one of %s", compatStr, allowedCompatStrings));
-  }
-
-  @VisibleForTesting
-  public static SparkCompat getCurrentSparkCompat() {
-    String sparkVersion = package$.MODULE$.SPARK_VERSION();
-    switch (sparkVersion.charAt(0)) {
-      case '3':
-        return SparkCompat.SPARK3_2_12;
-      case '2':
-        return SparkCompat.SPARK2_2_11;
-      default:
-        throw new IllegalStateException("Spark version " + sparkVersion + " is not known");
-    }
   }
 }
